yolov5/requirements.txt
# pip install -r requirements.txt

opencv-python>=4.1.2
torchvision>=0.8.1
numpy>=1.18.5
matplotlib>=3.2.2
kfp_pipeline_spec>=0.1.3.1
smbprotocol==1.4.0
h5py==2.10.0
kubernetes==11.0.0
kfp==1.6.4
onnxruntime==1.8.1
Pillow==8.3.1
secrets==1.0.2
tqdm>=4.41.0